added gitignore, created data folder to hide data from git, built support for rxapi
added gitignore
create datafolder to store data and hid from git
made config file
added functions to use in rxapi which should become our main script.
