+++ config_mdt.py
@@ -1,0 +1,8 @@
+#For MEPS year-specific variables
+meps_year = '18'
+
+#Age range: if user wants a drug distribution by age, Y, else N
+age_range_distrib = 'Y'
+
+#State: if user wants a distribution by state, Y, else N
+state_distrib = 'Y'
