+++ rxnorm_data_reader.py
@@ -11,19 +11,19 @@ z = zip_downloader('https://download.nlm.nih.gov/rxnorm/RxNorm_full_prescribe_cu
 
 #moves RxNorm files to sqlite database by reading as dataframes
 col_names = ['RXCUI','LAT','TS','LUI','STT','SUI','ISPREF','RXAUI','SAUI','SCUI','SDUI','SAB','TTY','CODE','STR','SRL','SUPPRESS','CVF','test']
-RXNCONSO = pd.read_csv(z.open('rrf/RXNCONSO.RRF'),sep='|',header=None,dtype=object,names=col_names)
-sql_create_table('RXNCONSO',RXNCONSO)
-del RXNCONSO
+rxnconso = pd.read_csv(z.open('rrf/RXNCONSO.RRF'),sep='|',header=None,dtype=object,names=col_names)
+sql_create_table('rxnconso',rxnconso)
+del rxnconso
 
 col_names = ['RXCUI1','RXAUI1','STYPE1','REL','RXCUI2','RXAUI2','STYPE2','RELA','RUI','SRUI','SAB','SL','DIR','RG','SUPPRESS','CVF','test']
-RXNREL = pd.read_csv(z.open('rrf/RXNREL.RRF'),sep='|',dtype=object,header=None,names=col_names)
-sql_create_table('RXNREL',RXNREL)
-del RXNREL
+rxnrel = pd.read_csv(z.open('rrf/RXNREL.RRF'),sep='|',dtype=object,header=None,names=col_names)
+sql_create_table('rxnrel',rxnrel)
+del rxnrel
 
 col_names = ['RXCUI','LUI','SUI','RXAUI','STYPE','CODE','ATUI','SATUI','ATN','SAB','ATV','SUPPRESS','CVF','test']
-RXNSAT = pd.read_csv(z.open('rrf/RXNSAT.RRF'),sep='|',dtype=object,header=None,names=col_names)
-sql_create_table('RXNSAT',RXNSAT)
-del RXNSAT
+rxnsat = pd.read_csv(z.open('rrf/RXNSAT.RRF'),sep='|',dtype=object,header=None,names=col_names)
+sql_create_table('rxnsat',rxnsat)
+del rxnsat 
 
 #deletes RxNorm ZIP
 del z
@@ -38,13 +38,13 @@ del rxcui_ndc
 
 
 #TEST!!!!!!!!!!!!!!!! reads record count from created database
-RXNCONSO = db_query("Select count(*) AS records from RXNCONSO limit 1")
+RXNCONSO = db_query("Select count(*) AS records from rxnconso limit 1")
 print('DB table RXNCONSO has {0} records'.format(RXNCONSO['records'].iloc[0]))
 
-RXNREL = db_query("Select count(*) AS records from RXNREL limit 1")
+RXNREL = db_query("Select count(*) AS records from rxnrel limit 1")
 print('DB table RXNREL has {0} records'.format(RXNREL['records'].iloc[0]))
 
-RXNSAT = db_query("Select count(*) AS records from RXNSAT limit 1")
+RXNSAT = db_query("Select count(*) AS records from rxnsat limit 1")
 print('DB table RXNSAT has {0} records'.format(RXNSAT['records'].iloc[0]))
 
 rxcui_ndc = db_query("Select count(*) AS records from rxcui_ndc limit 1")
